Refactor camera wizards and introduce probe positioning functionality
- Updated the camera wizards to include a new probe positioning wizard, allowing users to specify the location of probes or laser spots in the camera image.
- Added new pages for camera distortion correction and probe positioning, enhancing the user interface for better interaction.
- Removed the obsolete camerawizards.py file and organized the code into separate modules for clarity and maintainability.
- Improved documentation and user guidance within the wizard pages to assist users in positioning probes accurately.
- Fixes #58 with correct positioning of the marker when the objective lens selection changes
