+++ laserstudio/widgets/camerawizards/__init__.py
@@ -1,0 +1,10 @@
+from ...instruments.instruments import ProbeInstrument, LaserInstrument
+from .distortion import CameraDistortionWizard
+from .probeposition import ProbesPositionWizard
+
+__all__ = [
+    "ProbeInstrument",
+    "LaserInstrument",
+    "CameraDistortionWizard",
+    "ProbesPositionWizard",
+]
